basic: test 003 basic networking - with ovirtlib
